cmake: fix error for CMake before v3.28
On some distributions with non-latest CMake, it will fail on configuration due to its incapability of substition from unset variable to an empty string. To fix this, we need to add double quotes between variables that will be tested by EQUAL, LESS, GREATER, and other related keywords. For better compatibility, use STREQUAL instead of EQUAL if possible.

Signed-off-by: Athaariq Ardhiansyah <foss@athaariq.my.id>